requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
sqlalchemy==2.0.49
alembic==1.18.4
pydantic-settings==2.13.1
PyJWT==2.12.1
email-validator==2.3.0
python-multipart==0.0.24
pytest==9.0.3
httpx==0.28.1
jinja2==3.1.6
requests==2.32.5
ultralytics==8.3.1
